+++ HashTable/HashTable.h
@@ -10,16 +10,16 @@ class HashTable {
         unsigned int numOfBuckets;
 
     protected:
-        uint32_t HashToBucket(const char* string);
-        uint32_t HashToBucket(const uint32_t& hashValue);
+        uint64_t HashToBucket(const char* string);
+        uint64_t HashToBucket(const uint64_t& hashValue);
     public:
-        static uint32_t Hash(const char* stringData);
-        static uint32_t Hash(const uint32_t& integerData);
+        static uint64_t Hash(const char* stringData);
+        static uint64_t Hash(const uint64_t& integerData);
         explicit HashTable(const unsigned int& numOfBuckets);
         virtual ~HashTable();
-        uint32_t Insert(const char* input);
-        void Insert(const uint32_t& input, const char* inputString);
-        const char* GetStringBySecondaryHashTableKey(const uint32_t &primaryHashKey, const uint32_t& secondaryHashKey);
+        uint64_t Insert(const char* input);
+        void Insert(const uint64_t& input, const char* inputString);
+        const char* GetStringBySecondaryHashTableKey(const uint64_t &primaryHashKey, const uint64_t& secondaryHashKey);
 
         bool ContainsPrimaryHashTableKey(const char *inputString);
 };
